requirements.txt
opencv-python
imutils
pdf2image
numpy
pandas
requests
Pillow
PassportEye
boto3

